# docs/api/codex-bridge.openapi.yaml
openapi: 3.1.0
info:
  title: CodexBridge Mobile API
  summary: Contract-first HTTP API consumed by CodexBridgeMobile.
  version: 1.8.0
  description: |
    Canonical contract for the public HTTP API that `EDortta/CodexBridgeMobile`
    consumes. For that API this document is the **source of truth**: an
    endpoint under `/api` that is not here is not public, and a field that is
    not here is not part of the contract.

    This is the **only** OpenAPI description of this gateway. FastAPI's
    auto-generated document and its UIs (`/openapi.json`, `/docs`, `/redoc`) are
    switched off in `gateway/app/main.py`: they described the internal MCP and
    OAuth surfaces and carried none of the rules below, and serving them put two
    public descriptions of one gateway on the wire — with the reachable one not
    being the canonical one. There is no second document to reconcile against.

    ## What this document does not cover

    The MCP transport (`POST /mcp`), the OAuth authorization endpoints, the
    Prometheus scrape endpoint (`GET /metrics`) and the reverse executor
    WebSocket are **internal or third-party-facing** surfaces with their own
    contracts. Each one is listed in `x-contract-excluded-paths` below with the
    reason it is out of scope — see `docs/api/README.md` §"Contract scope".

    ## Versioning

    The public namespace is `/api/v1`. Compatibility, deprecation and the rules
    that force a new namespace are specified in `docs/api/README.md`. The
...

...

...

...
servers:
...
paths:
...
  /api/v1/events/stream:
    get:
      operationId: streamEvents
      tags: [events]
      summary: Live event stream (Server-Sent Events)
      description: |
        A `text/event-stream` body that stays open and delivers events as they
        are recorded. SSE rather than WebSocket for three reasons, stated so a
        client author does not have to guess: it rides the `Authorization:
        Bearer` scheme every other endpoint here uses (a browser or mobile
        WebSocket cannot set that header on the handshake, and a token in the
        URL is forbidden); `Last-Event-ID` puts exact resume in the transport
        rather than in an application protocol on top; and nothing in this
        feature needs a reverse channel.

        ## Frames

        Only entity events carry `id:`. A control frame deliberately does not,
        because `Last-Event-ID` must never advance past an event the client
        actually received.

        ```
        event: stream.open        (no id) the stream is open, from this position
        event: stream.gap         (no id) resume is not continuous — re-read
        event: <MobileEventType>  id: <n>  the events themselves, data is MobileEvent
        : keep-alive              a comment, so a proxy does not time out an idle stream
        event: stream.closed      (no id) why the server ended it
        ```

        `stream.closed` carries a `reason`: `max_duration` (the server bounds
        how long one connection lives — reconnect), `unauthenticated` (the
        credential stopped being usable) or `disconnected`. Every ending is a
        reconnect with `Last-Event-ID`, which is exact.

        ## Resume

        Send `Last-Event-ID` (an `EventSource` does this by itself) or `after`.
        The header wins when both are present: it is what a reconnecting client
        sends automatically and it is the more recent of the two by
        construction. A malformed `Last-Event-ID` is treated as absent rather
        than rejected — the user agent sets it, not the application.

        Resume is `id > position`, so an event cannot be delivered twice or
        skipped. If the position can no longer be continued from, the stream
        opens with `stream.gap` **before** delivering anything, so a client
        never acts on a partial view believing it was continuous.

        ## Authorization is re-checked while the stream runs

        A stream opened this morning and still open this afternoon authorized
        once. This one re-resolves the bearer token on every poll: a revoked
        token, an expired token, a disabled account and a project removed from
        the actor's `allowedProjects` all take effect within one poll interval,
        and the stream ends with `stream.closed` and `reason:
        unauthenticated`. A client must not treat an open connection as proof
        it is still authorized.

        ## What never appears here

        Authentication and security events. Sign-in, sign-in failure and
        credential revocation are recorded in the same audit log these events
        are derived from, and they are excluded by construction — streaming
        them would tell any token holder when an operator signs in. Neither do
        events whose project cannot be determined: they are delivered to
        nobody, administrators included.

        Preferences set through `PUT /api/v1/notifications/preferences` do
        **not** filter this stream. Narrow a connection with `?type=` instead.

        ## Limits

        The gateway bounds how many streams it holds open at once and answers
        `503` with `Retry-After` when it is at that ceiling; reconnect with
        `Last-Event-ID` and nothing is lost. `?type=` is validated before the
        first byte, because once an event-stream body has started there is no
        status code left to change.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventsAfter'
        - $ref: '#/components/parameters/EventProject'
        - $ref: '#/components/parameters/EventType'
        - name: Last-Event-ID
          in: header
          required: false
          description: |
            The last event id this client received, sent automatically by an
            `EventSource` on reconnect. Wins over `after` when both are given.
            A value that is not a non-negative integer is ignored.
          schema:
            type: string
      responses:
        '200':
          description: |
            An open event stream. The body is `text/event-stream` and does not
            end until the server closes it; it is not JSON and must be read
            frame by frame.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  A sequence of SSE frames. The `data:` field of an entity
                  frame is a `MobileEvent`; the `data:` field of a control
                  frame is a `StreamControlFrame`. Described as a string
                  because OpenAPI has no way to type a frame sequence — the
                  two schemas named in `x-frame-payloads` are what a client
                  parses each `data:` into.
              # OpenAPI's content model types a *body*, and this body is a
              # sequence of framed payloads of two different shapes. There is no
              # standard place to say that, so the mapping is a vendor
              # extension rather than prose: a generator can read it, and the
              # contract gate sees the two schemas as referenced instead of
              # reporting them as declared-and-unused, which is what they would
              # be if this said the same thing in a sentence.
              x-frame-payloads:
                entity:
                  description: The `data:` of a frame whose `event:` is a `MobileEventType`.
                  schema:
                    $ref: '#/components/schemas/MobileEvent'
                control:
                  description: The `data:` of a `stream.open`, `stream.gap` or `stream.closed` frame.
                  schema:
                    $ref: '#/components/schemas/StreamControlFrame'
              examples:
                openThenOneEvent:
                  summary: Opening, then one decision event
                  value: |
                    event: stream.open
                    data: {"type":"stream.open","from":0,"at":"2026-08-26T14:32:07Z"}

                    id: 41
                    event: decision.requested
                    data: {"id":41,"type":"decision.requested","projectId":"codexbridge-demo","entity":{"kind":"decision","id":"9f0d6b2c-2f5a-4c2b-9d5e-1a7c3f8e4b10"},"action":"requested","at":"2026-08-26T14:32:09Z","summary":"A sensitive session is waiting for a decision.","state":"awaiting_approval"}

                    : keep-alive

                    event: stream.closed
                    data: {"type":"stream.closed","reason":"max_duration","at":"2026-08-26T14:47:07Z"}
        '400': {$ref: '#/components/responses/ValidationFailed'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/PermissionDenied'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503':
          description: |
            The gateway is already holding as many event streams open as it
            will. Retry after the interval in `Retry-After`, with
            `Last-Event-ID`; nothing is lost by waiting.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: dependency_unavailable
                message: Too many event streams are open on this gateway. Retry shortly.
                requestId: '5a3f1c88-90b1-4e2a-8a6f-2f9c0d1e7b43'
                retryable: true
  /api/v1/notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      summary: This actor's notification-subscription preferences
      description: |
        Always the calling actor's own. There is no `userId` parameter and no
        administrator override: a preference document is personal data and an
        endpoint that could read another account's would be a disclosure with
        no product behind it.

        Returns the defaults — an empty `eventTypes` and `pushEnabled: false`,
        with `updatedAt: null` — when the actor has never saved preferences, so
        a client can tell "never touched" from "explicitly set to the
        defaults".
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The actor's preferences.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/PermissionDenied'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
    put:
      operationId: putNotificationPreferences
      tags: [notifications]
      summary: Replace this actor's notification-subscription preferences
      description: |
        A whole-document replacement, idempotent by construction: the same body
        sent twice produces the same document, which is why there is no
        `Idempotency-Key` and no `If-Match` here. The only writer of a row is
        the actor it belongs to, so there is no concurrent third party for an
        optimistic check to protect against.

        **These preferences are a hook for a push transport this build does not
        have.** `pushDeliveryAvailable` is `false` and nothing reads these rows
        to decide delivery. They are stored so the choice survives a reinstall
        and so a later push integration has something to read.

        **They do not filter `GET /api/v1/events/stream`.** A client that
        opened the stream asked for the stream; withholding events from it
        because of a preference set on another device is how a phone silently
        misses the decision its operator was waiting for. Narrow a live
        connection with that endpoint's `?type=` instead.

        Requires the `notifications.manage` permission
        (`codexbridge.notifications.manage`), which is separate from the read
        scope on purpose: an operator may want a phone that watches the event
        stream without being able to rewrite what the account is notified
        about.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferencesRequest'
      responses:
        '200':
          description: The stored preferences.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '400': {$ref: '#/components/responses/ValidationFailed'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/PermissionDenied'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  headers:
    XRequestId:
      description: |
        Identifier for this one request, echoed on every response including
        errors. A client may send it to tie its own log to the server's; a value
        that is not a valid `Id` is replaced rather than echoed, because this
        header is written into response headers and log lines.
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    RetryAfter:
      description: |
        Seconds to wait before repeating the request. Sent with `429` and with
        `503`. A client that ignores it and retries immediately will be limited
        again — treat it as the earliest useful retry, not a suggestion.
      schema:
        type: integer
        minimum: 0
...
    CacheControl:
      description: |
        `no-store` on every response that carries a credential or an
        authorization decision. A client must not write these bodies to disk,
        and an intermediary must not hold them.
      schema:
        type: string
      example: 'no-store'
  parameters:
...
    EventsAfter:
      name: after
      in: query
      required: false
      description: |
        Deliver only events with an `id` strictly greater than this. The
        resume position for both event transports, and deliberately the same
        integer the stream puts in `id:` — see `GET /api/v1/events`.

        On the stream, `Last-Event-ID` wins over this when both are sent.
      schema:
        type: integer
        format: int64
        minimum: 0
      examples:
        resumeFromFortyOne:
          value: 41
    EventProject:
      name: project
      in: query
      required: false
      description: |
        Repeatable. Restricts to these projects, **intersected** with what the
        caller may see — naming a project outside the actor's
        `allowedProjects` narrows to nothing rather than widening.
      schema:
        type: array
        items:
          $ref: '#/components/schemas/ProjectId'
    EventType:
      name: type
      in: query
      required: false
      description: |
        Repeatable. Restricts to these event types. A value outside
        `MobileEventType` is a `400`; a value that is a declared type this
        build never emits (`artifact.*`, `androidBuild.*`) is accepted and
        matches nothing, which is what makes those values usable by a client
        written ahead of the build that emits them.
      schema:
        type: array
        items:
          $ref: '#/components/schemas/MobileEventType'
  responses:
    ValidationFailed:
      description: The request was malformed or failed validation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: |
        No usable credential was presented. Absent, malformed, unknown, expired,
        revoked, or belonging to an account that has since been disabled or
        removed — one status, one code, one message for all of them, because
        telling them apart tells a caller holding a token they were never given
        whether it was ever real.

        This is the branch a client answers by signing in again. `403` is the
        other one.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PermissionDenied:
      description: |
        The actor is authenticated but not permitted. Returned in preference to
        `404` only when the resource's existence is already known to the caller.

        Signing in again does not help — the credential is fine, the permission
        is not. On `/api/v1` this comes from the per-action guard and from
        nowhere else, which is why operations with no action to guard (such as
        `getCurrentActor`) do not declare it.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Too many requests. `retryable` is true; honour `Retry-After`.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: |
        Unexpected server failure. The body carries only a `requestId`; the
        detail stays in the server log, because raw driver errors and stack
        traces leak schema, filesystem paths and sometimes credentials.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque access token, presented as `Authorization: Bearer <token>`.

        Two issuers, one credential store. `POST /api/v1/auth/sign-in` mints
        tokens for CodexBridgeMobile, and the OAuth 2.1 + PKCE authorization
        endpoints mint them for ChatGPT's MCP client. Both land in the same
        table, so `POST /api/v1/auth/revoke` takes effect on both surfaces and
        an expired or revoked token is refused everywhere.

        A token is opaque: it carries no readable claims and must not be parsed.
        Read `expiresIn` from the response that issued it, and refresh before it
        elapses rather than waiting for a `401`.

        Every rejection is `401` with code `unauthenticated` and the same
        message — absent, unknown, expired and revoked are not distinguished,
        because doing so tells a caller holding a token they were never given
        whether it was ever real. A client that receives `401` should refresh
        once, and sign in again if that also fails.
  schemas:
...
    ProjectId:
      allOf:
        - $ref: '#/components/schemas/Id'
      description: Identifier of a project registered in the gateway registry.
      examples:
        - 'codexbridge-demo'
...
    Error:
      type: object
      description: |
        The single error envelope returned by every endpoint in this contract
        for every non-2xx response. An endpoint that returns anything else is a
        contract violation, not a variant.
      required:
        - code
        - message
        - requestId
        - retryable
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          maxLength: 1024
          description: |
            Human-readable summary, safe to log. It is written for an operator,
            not for an end user, and it is not localized. Clients must not parse
            it and must not assume it is stable across releases.
        requestId:
          $ref: '#/components/schemas/Id'
          description: |
            Identifier tying this response to the server-side log record for
            **this one request**. Always present, including on `internal_error`,
            so an operator can find the failure from a mobile screenshot alone.

            Named `requestId`, not `correlationId`, on purpose: `correlation_id`
            already exists in this codebase as a column on `TaskModel`, scoped to
            the whole lifetime of a task and shared by the executor protocol.
            Populating this field from that column would make every failure on a
            given task report the same identifier, defeating the only reason the
            field exists — while the response stayed schema-valid, so no test
            would catch it. The two are different identifiers and must not be
            merged. When a response concerns a task, expose that task's
            correlation id as its own named field instead.
        retryable:
          type: boolean
          description: |
            Whether repeating the identical request may succeed later without
            the caller changing anything. `true` on transient failures
            (`rate_limited`, `dependency_unavailable`); `false` on failures that
            require caller action (`validation_failed`, `permission_denied`).

            This is advice about the failure, not permission to retry blindly:
            a client retrying a non-idempotent write still needs the idempotency
            rules that issue #12 defines.
        details:
          type: array
          description: |
            Per-problem breakdown. Present on `validation_failed`; may be
            omitted entirely on other codes. An empty array and an absent array
            mean the same thing.
          items:
            $ref: '#/components/schemas/ErrorDetail'
      examples:
        - code: validation_failed
          message: 'Request body failed validation.'
          requestId: '5a3f1c88-90b1-4e2a-8a6f-2f9c0d1e7b43'
          retryable: false
          details:
            - field: /timeoutSeconds
              code: above_maximum
              message: 'timeoutSeconds must not exceed the project limit of 3600.'
        - code: permission_denied
          message: 'The authenticated actor may not read this project.'
          requestId: 'c1d2e3f4-a5b6-4708-9a0b-1c2d3e4f5061'
          retryable: false
        - code: rate_limited
          message: 'Too many requests. Retry after the interval in Retry-After.'
          requestId: '7e8f9a0b-1c2d-4e3f-a405-162738495a6b'
          retryable: true
        - code: internal_error
          message: 'Unexpected server error. Report the requestId to the operator.'
          requestId: 'b4c5d6e7-f809-4a1b-8c2d-3e4f50617283'
          retryable: true
...
    MobileEvent:
      type: object
      description: |
        One change, in the shape both event transports deliver.

        **The summary is a whitelist, never the stored record.** The gateway's
        internal audit payload is not passed through: each event type names the
        few values it may read, free text is redacted and truncated on the way
        out, and anything nobody whitelisted does not leave the server. Treat
        `summary` as a notification line, not as data — fetch the entity for
        the authoritative state.
      required: [id, type, projectId, entity, action, at, summary]
      properties:
        id:
          type: integer
          format: int64
          description: |
            Monotonically increasing position in the event log, and the resume
            token: send it back as `Last-Event-ID` or `after`. Increasing but
            **not contiguous** — the log also holds records that are never
            delivered as events — so a client must not treat a skipped number
            as a lost event. Loss is reported explicitly by `EventGap`.
        type:
          $ref: '#/components/schemas/MobileEventType'
        projectId:
          $ref: '#/components/schemas/ProjectId'
        entity:
          $ref: '#/components/schemas/MobileEventEntity'
        action:
          type: string
          description: |
            The part of `type` after the dot (`created`, `state_changed`, …),
            so a client can branch coarsely without parsing the type string.
            Derived, never stored, so the two cannot disagree.
          examples:
            - state_changed
        at:
          $ref: '#/components/schemas/Timestamp'
        summary:
          type: string
          maxLength: 300
          description: One short human-readable line. See the note above.
          examples:
            - 'State changed to running.'
        state:
          type: string
          description: |
            The entity's state at the moment of the event, when the event has
            one. Omitted — not null — when it does not.
          examples:
            - running
        actorId:
          allOf:
            - $ref: '#/components/schemas/Id'
          description: |
            Opaque id of the actor who caused the change, when the event
            records one. Omitted when the change was not attributable to an
            actor (an executor result, an expiry sweep). Never an email
            address: the audit record holds one for several event kinds and it
            is deliberately not read here.
      examples:
        - id: 41
          type: session.state_changed
          projectId: codexbridge-demo
          entity:
            kind: session
            id: 9f0d6b2c-2f5a-4c2b-9d5e-1a7c3f8e4b10
          action: state_changed
          at: '2026-08-26T14:32:09Z'
          summary: 'State changed to running.'
          state: running
...
    StreamControlFrame:
      type: object
      description: |
        The `data:` payload of a `stream.open`, `stream.gap` or `stream.closed`
        frame on `GET /api/v1/events/stream`.

        These are **not** `MobileEventType` values and never will be: they carry
        no entity, no project and no resume position, and a client that switches
        exhaustively over the entity enum must not have to know them to compile.
        Control frames also carry no SSE `id:`, so they cannot advance
        `Last-Event-ID` past an event the client never received.
      required: [type]
      properties:
        type:
          type: string
          enum:
            - stream.open
            - stream.gap
            - stream.closed
        at:
          $ref: '#/components/schemas/Timestamp'
        from:
          type: integer
          format: int64
          description: On `stream.open` and `stream.gap`, the position the stream resumed from.
        reason:
          type: string
          description: |
            On `stream.gap`, an `EventGap.reason`. On `stream.closed`, why the
            server ended the stream: `max_duration` (the server bounds how long
            one connection lives), `unauthenticated` (the credential stopped
            being usable while the stream was open) or `disconnected`. Every
            one of them is a reconnect with `Last-Event-ID`.
        oldestAvailableId:
          type: [integer, 'null']
          format: int64
          description: On `stream.gap`, as `EventGap.oldestAvailableId`.
...